smart-stadium-os/backend/requirements.txt
fastapi>=0.103.1
uvicorn[standard]>=0.23.2
websockets>=11.0.3
pydantic>=2.7.0
pydantic-settings>=2.2.1
google-cloud-logging>=3.10.0
google-cloud-secret-manager>=2.20.0
firebase-admin>=6.5.0
pytest>=8.1.1
pytest-cov>=5.0.0
pytest-asyncio>=0.23.6
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-propagator-gcp>=1.6.0
